@@ -13,7 +13,7 @@ version = '8.0'
 if sys.argv[-1] == 'publish':
     try:
         import wheel
-        print("Wheel version: ", wheel.__version__)
+        print('Wheel version: ', wheel.__version__)
     except ImportError:
         print('Wheel library missing. Please run "pip install wheel"')
         sys.exit()
@@ -22,9 +22,9 @@ if sys.argv[-1] == 'publish':
     sys.exit()
 
 if sys.argv[-1] == 'tag':
-    print("Tagging the version on git:")
-    os.system("git tag -a %s -m 'version %s'" % (version, version))
-    os.system("git push --tags")
+    print('Tagging the version on git:')
+    os.system('git tag -a %s -m "version %s"' % (version, version))
+    os.system('git push --tags')
     sys.exit()
 
 readme = open('README.rst').read()
@@ -33,7 +33,7 @@ history = open('HISTORY.rst').read().replace('.. :changelog:', '')
 setup(
     name='django-elasticsearch-dsl',
     version=version,
-    python_requires=">=3.8",
+    python_requires='>=3.9',
     description="""Wrapper around elasticsearch-dsl-py for django models""",
     long_description=readme + '\n\n' + history,
     author='Sabricot',
@@ -46,25 +46,25 @@ setup(
         'elasticsearch-dsl>=8.9.0,<9.0.0',
         'six',
     ],
-    license="Apache Software License 2.0",
+    license='Apache Software License 2.0',
     zip_safe=False,
     keywords='django elasticsearch elasticsearch-dsl',
     classifiers=[
         'Development Status :: 3 - Alpha',
         'Framework :: Django',
-        'Framework :: Django :: 3.2',
-        'Framework :: Django :: 4.1',
         'Framework :: Django :: 4.2',
+        'Framework :: Django :: 5.1',
         'Intended Audience :: Developers',
         'License :: OSI Approved :: BSD License',
         'Natural Language :: English',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
         'Programming Language :: Python :: 3.11',
+        'Programming Language :: Python :: 3.12',
+        'Programming Language :: Python :: 3.13',
     ],
     extras_require={
-        'celery':  ["celery>=4.1.0"],
+        'celery':  ['celery>=4.1.0'],
     }
 )
