@@ -12,22 +12,22 @@ def is_es_online(connection_alias='default'):
         return es.ping()
 
 
-class ESTestCase(object):
-    _index_suffixe = '_ded_test'
+class ESTestCase:
+    _index_suffix = '_ded_test'
 
     def setUp(self):
         for doc in registry.get_documents():
-            doc._index._name += self._index_suffixe
+            doc._index._name += self._index_suffix
 
         for index in registry.get_indices():
-            index._name += self._index_suffixe
+            index._name += self._index_suffix
             index.delete(ignore=[404, 400])
             index.create()
 
-        super(ESTestCase, self).setUp()
+        super().setUp()
 
     def tearDown(self):
-        pattern = re.compile(self._index_suffixe + '$')
+        pattern = re.compile(re.escape(self._index_suffix) + '$')
 
         for index in registry.get_indices():
             index.delete(ignore=[404, 400])
@@ -36,4 +36,4 @@ class ESTestCase(object):
         for doc in registry.get_documents():
             doc._index._name = pattern.sub('', doc._index._name)
 
-        super(ESTestCase, self).tearDown()
+        super().tearDown()
