@@ -3,10 +3,7 @@ from __future__ import unicode_literals
 
 import django
 from django.db import models
-if django.VERSION < (4, 0):
-    from django.utils.translation import ugettext_lazy as _
-else:
-    from django.utils.translation import gettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from six import python_2_unicode_compatible
 
 
