@@ -1,11 +1,7 @@
 from unittest import TestCase
 
-import django
 from django.db.models.fields.files import FieldFile
-if django.VERSION < (4, 0):
-    from django.utils.translation import ugettext_lazy as _
-else:
-    from django.utils.translation import gettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from mock import Mock, NonCallableMock
 from six import string_types
 
@@ -17,7 +13,6 @@ from django_elasticsearch_dsl.fields import (BooleanField, ByteField, Completion
                                              ListField, LongField,
                                              NestedField, ObjectField, ScaledFloatField, ShortField, TextField
                                              )
-from tests import ES_MAJOR_VERSION
 
 
 class DEDFieldTestCase(TestCase):
@@ -80,7 +75,7 @@ class ObjectFieldTestCase(TestCase):
             'last_name': TextField()
         })
 
-        expected_type = 'string' if ES_MAJOR_VERSION == 2 else 'text'
+        expected_type = 'text'
 
         self.assertEqual({
             'type': 'object',
@@ -257,7 +252,7 @@ class NestedFieldTestCase(TestCase):
             'last_name': TextField()
         })
 
-        expected_type = 'string' if ES_MAJOR_VERSION == 2 else 'text'
+        expected_type = 'text'
 
         self.assertEqual({
             'type': 'nested',
@@ -397,7 +392,7 @@ class FileFieldTestCase(TestCase):
     def test_get_mapping(self):
         field = FileField()
 
-        expected_type = 'string' if ES_MAJOR_VERSION == 2 else 'text'
+        expected_type = 'text'
 
         self.assertEqual({
             'type': expected_type,
@@ -424,7 +419,7 @@ class TextFieldTestCase(TestCase):
     def test_get_mapping(self):
         field = TextField()
 
-        expected_type = 'string' if ES_MAJOR_VERSION == 2 else 'text'
+        expected_type = 'text'
 
         self.assertEqual({
             'type': expected_type,
@@ -434,13 +429,4 @@ class TextFieldTestCase(TestCase):
 class KeywordFieldTestCase(TestCase):
     def test_get_mapping(self):
         field = KeywordField()
-
-        if ES_MAJOR_VERSION == 2:
-            self.assertEqual({
-                'type': 'string',
-                'index': 'not_analyzed',
-            }, field.to_dict())
-        else:
-            self.assertEqual({
-                'type': 'keyword',
-            }, field.to_dict())
+        self.assertEqual({'type': 'keyword'}, field.to_dict())
