@@ -6,7 +6,7 @@ from .documents import Document  # noqa
 from .indices import Index  # noqa
 from .fields import *  # noqa
 
-__version__ = '7.1.1'
+__version__ = '8.1.0'
 
 
 def autodiscover():
